adds more flexibility to getDayWeather
now, when clicking on a day button, it will return the data from a specific hour, if an hour was selected previously, and if not, will return an approximate hour to the current one, or exactly the current one in the case of the current day.
The function approximateHour makes use of the date-fns dependency, so for the moment i'm importing it with a cdn link, until i build the app with webpack.